Gameクラスをファイルに切り出し
